fix not completed order tries to return quantities on delete

+++ backend/src/main/kotlin/org/napetrico/backend/features/orders/OrderService.kt
@@ -54,12 +54,14 @@ class OrderService(
         val user = userService.getCurrentUser()
         val order = getOrder(publicId, user)
 
-        val movements = order.movements
-        if (!movements.isEmpty())
-            movements.forEach {
-                updateQuantity(it, -it.quantity)
-                movementService.deleteMovement(it.publicId, user)
-            }
+        if (order.isCompleted) {
+            val movements = order.movements
+            if (!movements.isEmpty())
+                movements.forEach {
+                    updateQuantity(it, -it.quantity)
+                    movementService.deleteMovement(it.publicId, user)
+                }
+        }
 
         orderRepository.delete(order)
     }
